# replace spatial with per_activation mode for bn op when input is NCHW and H==1 and W==1 to improve perf