@zhangting2020 2021-07-01
性能差于torch的贴一下nvprof的输出？是仅统计了batch norm的kernel时间，还是所有的gpu时间？